Moved all the primitive type definition into script. Addess struct declaration problem in generate code.

+++ engine/script/test/case/foo.cpp
@@ -2,7 +2,7 @@
 
 //[[TypeInfo]]//
 struct Bar {
-	long a;
+	double a;
 	Foo * foo_ptr;
 	Foo &foo_ref;
 };
